+++ multishape.cpp
@@ -2,7 +2,7 @@
 
 multishape::~multishape() {}
 
-void multishape::save(std::vector<uint16_t>& instrs, std::vector<uint32_t>& uint32s, std::vector<float>& floats, std::vector<float>& x_coords, std::vector<float>& y_coords, std::vector<float>& z_coords) const {
+void multishape::save(std::vector<uint8_t>& instrs, std::vector<uint32_t>& uint32s, std::vector<float>& floats, std::vector<float>& x_coords, std::vector<float>& y_coords, std::vector<float>& z_coords) const {
   for (auto instr : instructions) {
     instrs.push_back(to_underlying(instr.instr));
     x_coords.insert(std::end(x_coords), std::begin(instr.x_coords), std::end(instr.x_coords));
@@ -129,7 +129,7 @@ void multishape::dump() {
   }
 }
 
-void multishape::load(drw_file& drw, parasitic_vector<uint16_t> &instructions, parasitic_vector<float>& x_coords, parasitic_vector<float>& y_coords, parasitic_vector<float>& z_coords,
+void multishape::load(drw_file& drw, parasitic_vector<uint8_t> &instructions, parasitic_vector<float>& x_coords, parasitic_vector<float>& y_coords, parasitic_vector<float>& z_coords,
   parasitic_vector<float>& floats, parasitic_vector<uint32_t>& uint32s, uint32_t& instr_index, uint32_t& current_uint32, uint32_t& current_x_coord,
   uint32_t& current_y_coord, uint32_t& current_z_coord, uint32_t& current_float) {
