+++ multishape.hpp
@@ -30,15 +30,15 @@ class multishape : public shape {
     vertices.push_back(y);
   }
   void dump();
-  void save(std::vector<uint16_t>& instructions, std::vector<uint32_t>& uint32s, std::vector<float>& floats, std::vector<float>& x_coords, std::vector<float>& y_coords, std::vector<float>& z_coords) const override;
-  void load(drw_file& drw, parasitic_vector<uint16_t> &instructions, parasitic_vector<float>& x_coords, parasitic_vector<float>& y_coords, parasitic_vector<float>& z_coords,
+  void save(std::vector<uint8_t>& instructions, std::vector<uint32_t>& uint32s, std::vector<float>& floats, std::vector<float>& x_coords, std::vector<float>& y_coords, std::vector<float>& z_coords) const override;
+  void load(drw_file& drw, parasitic_vector<uint8_t> &instructions, parasitic_vector<float>& x_coords, parasitic_vector<float>& y_coords, parasitic_vector<float>& z_coords,
             parasitic_vector<float>& floats, parasitic_vector<uint32_t>& uint32s, uint32_t& instr_index, uint32_t& current_uint32, uint32_t& current_x_coord,
             uint32_t& current_y_coord, uint32_t& current_z_coord, uint32_t& current_float);
 
 
   struct dispatch_inputs {
     multishape* shape;
-    parasitic_vector<uint16_t> &instructions;
+    parasitic_vector<uint8_t> &instructions;
     parasitic_vector<float> &x_coords;
     parasitic_vector<float> &y_coords;
     parasitic_vector<float> &z_coords;
